# Backend/MainService/openapi.yaml
openapi: 3.1.0
info:
  title: User Service API
  version: 1.0.0
  description: OpenAPI спецификация для основного сервиса
paths:
  /api/v1/main_service/projects/{projectId}/subscriptions:
    post:
      tags:
        - ProjectSubscription
      summary: Подписка на проект
      description: Подписывает текущего пользователя на указанный проект.
      operationId: subscribeToProject
      parameters:
        - name: projectId
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/XUserNameHeader'
      responses:
        '200':
          description: Подписка успешно выполнена
          content: { }
    delete:
      tags:
        - ProjectSubscription
      summary: Отписка от проекта
      description: Отписывает текущего пользователя от указанного проекта.
      operationId: unsubscribeFromProject
      parameters:
        - name: projectId
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/XUserNameHeader'
      responses:
        '204':
          description: Пользователь успешно отписан
          content: { }
    get:
      tags:
        - ProjectSubscription
      summary: Проверка подписки пользователя
      description: Проверяет, подписан ли текущий пользователь на проект.
      operationId: isUserSubscribed
      parameters:
        - name: projectId
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/XUserNameHeader'
      responses:
        '200':
          description: Статус подписки
          content:
            application/json:
              schema:
                type: boolean
                example: true
...
  /api/v1/main_service/projects/{projectId}/searches/employees:
    post:
      tags:
        - ProjectSearch
      summary: Создание записи поиска сотрудника
      operationId: createEmployeeSearch
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeSearchRequest'
      responses:
        '201':
          description: Запись создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeSearchFullResponse'
    get:
      tags:
        - ProjectSearch
      summary: Получение записей поиска сотрудников проекта
      operationId: getEmployeeSearches
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список записей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployeeSearchFullResponse'
  /api/v1/main_service/projects/{projectId}/searches/employees/{id}:
    put:
      tags:
        - ProjectSearch
      summary: Обновление записи поиска сотрудника
      operationId: updateEmployeeSearch
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeSearchRequest'
      responses:
        '200':
          description: Запись обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeSearchFullResponse'
    delete:
      tags:
        - ProjectSearch
      summary: Удаление записи поиска сотрудника
      operationId: deleteEmployeeSearch
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Запись удалена
          content: { }
  /api/v1/main_service/projects/{projectId}/searches/founders:
    post:
      tags:
        - ProjectSearch
      summary: Создание записи поиска сооснователя
      operationId: createFounderSearch
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FounderSearchRequest'
      responses:
        '201':
          description: Запись создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FounderSearchFullResponse'
    get:
      tags:
        - ProjectSearch
      summary: Получение записей поиска сооснователей проекта
      operationId: getFounderSearches
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список записей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FounderSearchFullResponse'
  /api/v1/main_service/projects/{projectId}/searches/founders/{id}:
    put:
      tags:
        - ProjectSearch
      summary: Обновление записи поиска сооснователя
      operationId: updateFounderSearch
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FounderSearchRequest'
      responses:
        '200':
          description: Запись обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FounderSearchFullResponse'
    delete:
      tags:
        - ProjectSearch
      summary: Удаление записи поиска сооснователя
      operationId: deleteFounderSearch
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Запись удалена
          content: { }
  /api/v1/main_service/projects/{projectId}/searches/investors:
    post:
      tags:
        - ProjectSearch
      summary: Создание записи поиска инвестора
      operationId: createInvestorSearch
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvestorSearchRequest'
      responses:
        '201':
          description: Запись создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvestorSearchFullResponse'
    get:
      tags:
        - ProjectSearch
      summary: Получение записей поиска инвесторов проекта
      operationId: getInvestorSearches
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список записей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvestorSearchFullResponse'
  /api/v1/main_service/projects/{projectId}/searches/investors/{id}:
    put:
      tags:
        - ProjectSearch
      summary: Обновление записи поиска инвестора
      operationId: updateInvestorSearch
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvestorSearchRequest'
      responses:
        '200':
          description: Запись обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvestorSearchFullResponse'
    delete:
      tags:
        - ProjectSearch
      summary: Удаление записи поиска инвестора
      operationId: deleteInvestorSearch
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Запись удалена
          content: { }
  /api/v1/main_service/projects/{projectId}/searches/mentors:
    post:
      tags:
        - ProjectSearch
      summary: Создание записи поиска ментора
      operationId: createMentorSearch
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MentorSearchRequest'
      responses:
        '201':
          description: Запись создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MentorSearchFullResponse'
    get:
      tags:
        - ProjectSearch
      summary: Получение записей поиска менторов проекта
      operationId: getMentorSearches
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список записей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MentorSearchFullResponse'
  /api/v1/main_service/projects/{projectId}/searches/mentors/{id}:
    put:
      tags:
        - ProjectSearch
      summary: Обновление записи поиска ментора
      operationId: updateMentorSearch
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MentorSearchRequest'
      responses:
        '200':
          description: Запись обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MentorSearchFullResponse'
    delete:
      tags:
        - ProjectSearch
      summary: Удаление записи поиска ментора
      operationId: deleteMentorSearch
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Запись удалена
          content: { }
...
components:
  schemas:
...
    EmployeeSearchRequest:
      type: object
      required:
        - specialization
        - about
      properties:
        specialization:
          type: string
          description: Название специализации
          minLength: 1
          example: "Backend Developer"
        about:
          type: string
          description: Описание требований
          example: "Ищем опытного backend разработчика"
    FounderSearchRequest:
      type: object
      required:
        - domain
        - about
      properties:
        domain:
          type: string
          description: Направление деятельности
          example: "FinTech"
        about:
          type: string
          description: Описание требований
          example: "Ищем сооснователя с опытом в финансах"
    InvestorSearchRequest:
      type: object
      required:
        - about
      properties:
        about:
          type: string
          description: Описание требований
          example: "Ищем инвестора для seed-раунда"
    MentorSearchRequest:
      type: object
      required:
        - domain
        - about
      properties:
        domain:
          type: string
          description: Направление деятельности
          example: "FinTech"
        about:
          type: string
          description: Описание требований
          example: "Ищем ментора с опытом в EdTech"
    EmployeeSearchFullResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор записи поиска сотрудника
          example: 1
        specialization:
          type: string
          description: Название специализации
          example: "Backend Developer"
        about:
          type: string
          description: Описание требований
          example: "Ищем опытного backend разработчика"
    FounderSearchFullResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор записи поиска сооснователя
          example: 1
        domain:
          type: string
          description: Направление деятельности
          example: "FinTech"
        about:
          type: string
          description: Описание требований
          example: "Ищем сооснователя с опытом в финансах"
    InvestorSearchFullResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор записи поиска инвестора
          example: 1
        about:
          type: string
          description: Описание требований
          example: "Ищем инвестора для seed-раунда"
    MentorSearchFullResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор записи поиска ментора
          example: 1
        domain:
          type: string
          description: Направление деятельности
          example: "EdTech"
        about:
          type: string
          description: Описание требований
          example: "Ищем ментора с опытом в EdTech"
...
  parameters:
    XUserNameHeader:
      name: X-User-Name
      in: header
      description: "Имя пользователя, проставляется Gateway"
      required: true
      schema:
        type: string